Rename LinkedIn.ipynb to Data Transformation using explode() & pivot()
To start off, we import the dataset using pandas and handle the comma-separated values. Then, we use the "explode()" function to split the values into individual rows. After renaming the columns, we use the "pivot()" function to create a table with Covid zones for each area on different dates. Finally, we write the output to an excel sheet.